Исправлены ошибки и добавлен вывод категорий при нажатии на первую категорию (корневую)

+++ src/main/java/ru/step/store/controllers/CategoryController.java
@@ -1,6 +1,10 @@
 package ru.step.store.controllers;
 
+import org.hibernate.collection.internal.PersistentSet;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import ru.step.store.models.Category;
+import ru.step.store.models.Role;
+import ru.step.store.models.User;
 import ru.step.store.repositories.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -17,18 +21,27 @@ public class CategoryController {
     private CategoryRepository categoryRepository;
 
     @GetMapping("/category/add")
-    public String getCategory() {
+    public String getCategory(@AuthenticationPrincipal User user) {
+        if (!((Role) user.getRoles().toArray()[0]).name().equals("ADMIN")) {
+            return "redirect:/";
+        }
         return "admin/category";
     }
 
     @GetMapping("/categories")
-    public String getCategories(Model model) {
+    public String getCategories(Model model, @AuthenticationPrincipal User user) {
+        if (!((Role) user.getRoles().toArray()[0]).name().equals("ADMIN")) {
+            return "redirect:/";
+        }
         model.addAttribute("categories", categoryRepository.findAll());
         return "admin/categories";
     }
 
     @PostMapping("/category/create")
-    public String createCategory(@RequestParam("name") String name, @RequestParam("categoryId") String categoryId) {
+    public String createCategory(@RequestParam("name") String name, @RequestParam("categoryId") String categoryId, @AuthenticationPrincipal User user) {
+        if (!((Role) user.getRoles().toArray()[0]).name().equals("ADMIN")) {
+            return "redirect:/";
+        }
         Category category = new Category();
         category.setName(name);
         if (categoryId.isEmpty()) {
@@ -41,7 +54,10 @@ public class CategoryController {
     }
 
     @GetMapping("/categories/{id}")
-    public String editCategory(@PathVariable("id") Long id, Model model) {
+    public String editCategory(@PathVariable("id") Long id, Model model, @AuthenticationPrincipal User user) {
+        if (!((Role) user.getRoles().toArray()[0]).name().equals("ADMIN")) {
+            return "redirect:/";
+        }
         model.addAttribute("category", categoryRepository.findById(id).get());
         return "admin/edit-category";
     }
@@ -49,7 +65,10 @@ public class CategoryController {
     @PostMapping("/category/{id}")
     public String updateCategory(@PathVariable("id") Long id,
                                  @RequestParam("categoryId") String categoryId,
-                                 @RequestParam("name") String name) {
+                                 @RequestParam("name") String name, @AuthenticationPrincipal User user) {
+        if (!((Role) user.getRoles().toArray()[0]).name().equals("ADMIN")) {
+            return "redirect:/";
+        }
         if (!categoryId.isEmpty()) {
             if (Long.parseLong(categoryId) == id) {
                 return "redirect:/categories";
@@ -67,7 +86,10 @@ public class CategoryController {
     }
 
     @GetMapping("/category/delete/{id}")
-    public String deleteCategory(@PathVariable("id") Long id) {
+    public String deleteCategory(@PathVariable("id") Long id, @AuthenticationPrincipal User user) {
+        if (!((Role) user.getRoles().toArray()[0]).name().equals("ADMIN")) {
+            return "redirect:/";
+        }
         categoryRepository.delete(categoryRepository.findById(id).get());
         return "redirect:/categories";
     }
